requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
flower==2.0.1
gunicorn==23.0.0
humanize==4.14.0
kombu==5.5.4
packaging==25.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
